+++ src/main/kotlin/compiler/ast/Node.kt
@@ -1,8 +1,8 @@
 package com.dlfsystems.compiler.ast
 
+import com.dlfsystems.server.Yegg
 import com.dlfsystems.compiler.Coder
 import com.dlfsystems.compiler.CompileException
-import io.viascom.nanoid.NanoId
 
 // A node in the syntax tree.
 
@@ -35,6 +35,6 @@ abstract class Node {
     open fun code(code: Coder) { }
 
     companion object {
-        fun makeID() = NodeID(NanoId.generateOptimized(8, "0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ", 61, 16))
+        fun makeID() = NodeID(Yegg.newID())
     }
 }
