--- src/main/kotlin/world/trait/Prop.kt
@@ -1,4 +1,0 @@
-package com.dlfsystems.world.trait
-
-class Prop {
-}
